Implement reviewers suggestions. Add documentation. Add dispatching mechanism to the test.
