+++ test/integration_test.exs
@@ -1,102 +1,9 @@
 defmodule Membrane.Agora.IntegrationTest do
   use ExUnit.Case
 
+  alias Membrane.Agora.Support.{SenderPipeline, ReceiverPipeline}
+
   import Membrane.Testing.Assertions
-
-  defmodule FramerateAsserter do
-    use Membrane.Filter
-
-    @tolerance Membrane.Time.milliseconds(100)
-
-    def_input_pad :input, accepted_format: _any, flow_control: :auto
-    def_output_pad :output, accepted_format: _any, flow_control: :auto
-
-    def_options framerate: []
-
-    @impl true
-    def handle_init(_ctx, opts) do
-      {num, den} = opts.framerate
-      interval = Membrane.Time.seconds(Ratio.new(den, num))
-      {[], %{interval: interval, last_timestamp: nil, how_many_frames: 0}}
-    end
-
-    @impl true
-    def handle_buffer(:input, buffer, _ctx, state) do
-      if state.last_timestamp do
-        diff = buffer.dts - state.last_timestamp
-        assert abs(diff - state.interval) < @tolerance
-
-        assert abs(state.interval * state.how_many_frames - buffer.dts) < 5 * @tolerance
-      end
-
-      {[buffer: {:output, buffer}],
-       %{state | last_timestamp: buffer.dts, how_many_frames: state.how_many_frames + 1}}
-    end
-  end
-
-  defmodule ReceiverPipeline do
-    use Membrane.Pipeline
-
-    @channel_name System.get_env("AGORA_CHANNEL_NAME", "")
-    @token System.get_env("AGORA_TOKEN", "")
-    @app_id System.get_env("AGORA_APP_ID", "")
-    @user_id System.get_env("AGORA_USER_ID", "0")
-
-    @impl true
-    def handle_init(_ctx, opts) do
-      spec = [
-        child(:source, %Membrane.Agora.Source{
-          channel_name: @channel_name,
-          token: @token,
-          app_id: @app_id,
-          user_id: @user_id
-        })
-        |> via_out(:video)
-        |> child(%FramerateAsserter{framerate: opts[:framerate]})
-        |> child(:video_sink, %Membrane.File.Sink{location: opts[:video]}),
-        get_child(:source)
-        |> via_out(:audio)
-        |> child(:audio_sink, %Membrane.File.Sink{location: opts[:audio]})
-      ]
-
-      {[spec: spec], %{}}
-    end
-  end
-
-  defmodule SenderPipeline do
-    use Membrane.Pipeline
-
-    @channel_name System.get_env("AGORA_CHANNEL_NAME", "")
-    @token System.get_env("AGORA_TOKEN", "")
-    @app_id System.get_env("AGORA_APP_ID", "")
-    @user_id System.get_env("AGORA_USER_ID", "0")
-
-    @impl true
-    def handle_init(_ctx, opts) do
-      spec =
-        [
-          child(%Membrane.File.Source{location: opts[:video]})
-          |> child(%Membrane.H264.Parser{
-            generate_best_effort_timestamps: %{framerate: opts[:framerate]}
-          })
-          |> child(Membrane.Realtimer)
-          |> via_in(:video)
-          |> child(:sink, %Membrane.Agora.Sink{
-            channel_name: @channel_name,
-            token: @token,
-            app_id: @app_id,
-            user_id: @user_id
-          }),
-          child(%Membrane.File.Source{location: opts[:audio]})
-          |> child(Membrane.AAC.Parser)
-          |> child(Membrane.Realtimer)
-          |> via_in(:audio)
-          |> get_child(:sink)
-        ]
-
-      {[spec: spec], %{}}
-    end
-  end
 
   @tag :tmp_dir
   test "if the data is sent to Agora properly", %{tmp_dir: dir} do
